+++ sdk/appconfiguration/azure-appconfiguration/tests/test_azure_configuration_client_async.py
@@ -462,7 +462,7 @@ class AppConfigurationClientTest(AzureTestCase):
     def test_sync_tokens(self, client):
         new = FeatureFlagConfigurationSetting(
             'custom',
-            True,
+            enabled=True,
             filters = [
                 {
                     "name": "Microsoft.Percentage",
@@ -481,7 +481,7 @@ class AppConfigurationClientTest(AzureTestCase):
 
         new = FeatureFlagConfigurationSetting(
             'time_window',
-            True,
+            enabled=True,
             filters = [
                 {
                     u"name": TIME_WINDOW,
@@ -500,7 +500,7 @@ class AppConfigurationClientTest(AzureTestCase):
 
         new = FeatureFlagConfigurationSetting(
             "newflag",
-            True,
+            enabled=True,
             filters=[
                 {
                     "name": TARGETING,
@@ -540,7 +540,7 @@ class AppConfigurationClientTest(AzureTestCase):
 
     @app_config_decorator
     def test_config_setting_feature_flag(self, client):
-        feature_flag = FeatureFlagConfigurationSetting("test_feature", True)
+        feature_flag = FeatureFlagConfigurationSetting("test_feature", enabled=True)
         set_flag = client.set_configuration_setting(feature_flag)
 
         self._assert_same_keys(feature_flag, set_flag)
@@ -597,7 +597,7 @@ class AppConfigurationClientTest(AzureTestCase):
     def test_feature_filter_targeting(self, client):
         new = FeatureFlagConfigurationSetting(
             "newflag",
-            True,
+            enabled=True,
             filters=[
                 {
                     "name": TARGETING,
@@ -657,7 +657,7 @@ class AppConfigurationClientTest(AzureTestCase):
     def test_feature_filter_time_window(self, client):
         new = FeatureFlagConfigurationSetting(
             'time_window',
-            True,
+            enabled=True,
             filters=[
                 {
                     "name": TIME_WINDOW,
@@ -682,7 +682,7 @@ class AppConfigurationClientTest(AzureTestCase):
     def test_feature_filter_custom(self, client):
         new = FeatureFlagConfigurationSetting(
             'custom',
-            True,
+            enabled=True,
             filters=[
                 {
                     "name": PERCENTAGE,
@@ -707,7 +707,7 @@ class AppConfigurationClientTest(AzureTestCase):
     def test_feature_filter_multiple(self, client):
         new = FeatureFlagConfigurationSetting(
             'custom',
-            True,
+            enabled=True,
             filters=[
                 {
                     "name": PERCENTAGE,